Add support for declared functions
Summary:
This diff takes the providers reported by the provider analysis and intersects them together to get the type for declared functions.

When we see a read of a declared function in the name_resolver we emit a special `DeclaredFunction` Val.t that lets us know to perform this lookup.

`Passed: 397, Failed: 71, Skipped: 287, Errored: 0`

Changelog: [Internal]

Reviewed By: panagosg7

Differential Revision: D34399552

fbshipit-source-id: 8eccbcbb21d3e7c2774326ab3d84e01498199c50