Fix heap refinements bug
Summary:
I was accidentally unifying the unrefined type with the heap refinement slot instead of the refined type because I mistakenly believed that the first type given was the type of the object itself, not the heap refinement. This diff renames those variables and fixes that bug.

`Passed: 398, Failed: 70, Skipped: 287, Errored: 0`

Changelog: [Internal]

Reviewed By: panagosg7

Differential Revision: D34401544

fbshipit-source-id: 4e555f66f3d65cb20ef39fea4d613d60c14808ce