Encode self-referential name in ThisClassT
Summary:
In order to implement capture avoiding substitution, we need to be able to alpha-rename type abstractions. We can do so already with `PolyT`s, but type parameters are also abstracted by `ThisClassT`, which currently binds "this". In order to allow us to alpha-rename "this", we add a new entry to `ThisClassT`'s tuple, which is its local name for the self-reference. This will always start out as "this," but may be alpha-renamed later.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D34018355

fbshipit-source-id: f1395c19bd461cd496607995832d1db8ff8b1133