Use GenericTs in normalization
Summary:
Continuing from the previous diff, this diff replaces `BoundT`s in type matching with `GenericT`s. This change shouldn't have any real ramifications for how Flow's typechecking works, but is necessary in order to fully remove the BoundT type. In order to generate `GenericT`s, we need to be able to construct aloc IDs, and in order to do that, we need to thread through the context into type matching.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D34034741

fbshipit-source-id: 568406df8a789197bb0d69662af0e33f6c753cb1