Don't look for escaped generics in annotated parameters and declarations
Summary:
This tweaks the escaped generics analysis in preparation for changes to GenericTs later on. We'll no longer consider function parameters to be in-scope for function or class type parameters, which means that we'll error if we detect a parameter with a lower bound of a GenericT. However, we don't do this inspection on parameters that are annotated, since generics can't escape in that case. The result is that a parameter annotated with a generic is not an error, but one inferred to be a generic is.

This permutes some error messages, but should not change whether or not a specific escape is detected.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D33399734

fbshipit-source-id: 63f43f757fe9e5abc6af883b0ce8b481dd485e20