Replace BoundT with GenericT
Summary:
This diff changes the behavior of checking generic functions and classes to rely entirely on `GenericT` types, rather than both `GenericT` and `BoundT`. The `GenericT` type now encapsulates both the behavior of `GenericT`  (checked in constraints, represents a range of types) and `BoundT` (substituted away by instantiations). Later diffs will remove `BoundT` completely.

This change is motivated by work on the fully-resolved environment. In the resolved environment, each variable binding (whether a type or a term) is uniquely mapped by binding location to a single resolved type. This approach isn't really compatible with the way that tparams are represented currently. Consider, for example, the definition

```function f<T>(x: T): T { var y: T; ... }
```
Here, there are four bindings that are created in the fully resolved environment: `f`, `T`, `x`, and `y`. However, currently "`T`" refers to two different things in differnent places: in `var y: T`, `T` refers to a `GenericT`, and in the return type of `f`, it refers to a `BoundT`. In the parameter `(x: T)`, it actually refers to both `BoundT` and `GenericT` at different times. By collapsing both the types to a single `GenericT` that works in both contexts, we can bring generics back into alignment with the fully resolved environment.

In order to combine `BoundT` and `GenericT`, `GenericT` needs to take on the behavior of `BoundT`. The central behavior here is that of substitutability, so `subst.ml` now knows to replace `GenericT`s whose name match the type being substituted away. There are a small handful of other places where we also change matches on `BoundT` to include `GenericT` as well.

We leave changing the presence of `BoundT` in type-at-pos and merging for later diffs.

-----
## Coverage

This change induces some changes to the coverage computation, which have only minor effects here but which will be multiplied in D34018353 (which applies this change to merging). I'm documenting this here since the *source* of the change is in this diff, even though the consequences only become apparent there. Currently, a `BoundT` is always considered to be covered ("controversially" according to a comment), even if its upper bound is uncovered. On the other hand, a GenericT is uncovered if its bound is uncovered. This means that previous to this diff, a case like`function f<X: any>(x: X) { x }` would consider the parameter `x` to be covered by the expression `x` in the body to be uncovered. By changing everything to be GenericTs, both now are considered uncovered.

----
## Free variable capture
Finally, this change introduces (or rather, surfaces) a bug with how substitution works. Flow's substitution currently doesn't perform capture avoidance. In the lambda calculus, capture-avoiding substitution is necessary whenever free variables may appear in the terms that that are replacing variables, in order to ensure that those free variables aren't captured by abstractions. For example,
```
(λx.y)[y := x]
```
Here, `x` is a free variable when it replaces `y` inside the target of substitution. But if we naively applied this substitution directly, we would end up with
```
λx.x
```
where `x` has been "captured" by the lambda: previously it was free and presumably referred to some other binding for `x` in scope; now it refers to the `x` bound by the λ. In order to safely perform this substitution, the classic approach is to alpha-rename the binding `x` to another, fresh name, e.g.
```
λz.x
```
Bringing this back to this diff: Flow currently (both before and after this change) does not perform capture-avoiding substitution. Originally, that didn't matter, because there was a distinction between bound and free variables--even though we could end up with a term like `λx.x` through substitution, we would know that, if the `x` in the body is a `GenericT` then it refers to a free var, while if it's a `BoundT` it refers to the bound var. We are erasing this distinction now, so suddenly variable capture has become a problem. This diff adds a test case that documents one way that capturing has become a problem. We implement capture-avoiding substitution later in D34018350; this diff should not land until that one is ready to land.

Changelog: [coverage] Changed the way Flow computes coverage for type parameters with uncovered upper bounds. Previously, a type parameter was always considered covered when it appeared in a function signature, class declaration, or type alias, even if the type parameter's upper bound was `any`. With this change, such type parameters are now considered to be uncovered. This results in Flow now computing a lower coverage percentage on modules where this pattern is present or that import types from other modules that use this pattern.

Reviewed By: panagosg7

Differential Revision: D33394405

fbshipit-source-id: 967d1d1c65959c2ac034582ea374c4bd65022117