Use GenericTs in type_sig_merge
Summary:
This diff removes the final place where BoundTs are generated, in `type_sig_merge`. This turns out to be a substantial change, because the beauty of `BoundT`s is that they're very flat and don't say that much on their own, while GenericTs have substantially more structure to them by encoding their own bounds. In order to know what bound a GenericT has, we need to thread through an environment-like structure in type_sig_merge, which records the GenericT types for in-scope type parameters.

The GenericTs themselves are straightforward to generate for regular type params, but for `this`, we need to introduce additional structure in the form of an unresolved avar. The upper bound for the `this` GenericT can then be the unresolved var when we're merging the type of the class, and once we have that type we can resolve the avar to it, tying the knot (just as we do during checking).

Reviewed By: samwgoldman

Differential Revision: D34018353

fbshipit-source-id: ee39355467ac8ab3a9c929b80e666d01e6d294f1