Implement capture-avoiding substitution
Summary:
Here we finally implement capture-avoiding substitution. When performing a substitution, we compute the free vars within the expressions that are replacing variables in the target (by which I mean: given a substitution like `e1[x:=e2]`, the FV's inside `e2`). Then, when we substitute through a binding (a PolyT or ThisClassT) where the abstracted variable names match any of those free variables, we rename the binding to avoid conflicts. We then have to add the renaming to the set of substitutions in order to make sure that references to the renamed variable are themselves properly renamed.

Tricky bits: we need to make sure that our process for renaming variables is deterministic so that we can still use the constraint cache--that's why the `Subst_name.Id` generation doesn't just use a random or globally-incrementing int. And, we need to be careful about changing PolyT ids. If we change the ID any time a substitution happens, we'll have performance problems because we'll miss the fast paths. So we use the basic lambda calculus insight that expressions are equivalent under a safe alpha renaming, and so we only change a PolyT id if a *non-renaming* substitution occurs within its body.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D34018350

fbshipit-source-id: c74363894c0cc9047ccc82a41df8841d6c706fd8