Use new subst_name type for possible subjects of substitution rather than strings
Summary:
In order to fix the free variable capture problem discussed in D33394405, we'll need to start renaming type parameters. We want to do this renaming smartly, and we want to be able to normalize and re-display names with the original, user-visible name. So this diff moves from using strings to represent tparam names and subsitution to using an new data structure that can either be a regular `Name`, an `Id` which is the result of renaming, or a `Synthetic` name which cannot be renamed or substituted.

Currently, `Id`s are never generated, because we're not yet implementing capture avoiding substitution. In most cases, `Name`s are all we use, and they behave identically to strings for substitution. `Synthetic` names are used when converting from more complex generic encodings such as spreads, where a single GenericT can represent more than one source-level generic. Such GenericTs should never be the targets of substitution, because they only arise through constraint solving (rather than in signatures where they can be replaced in substitutions). We enforce this by erroring if a Synthetic name is substituted.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D34018351

fbshipit-source-id: cce684f7fd068d72f02077eb3043545966c939b4