Remove unnecessary substitutions
Summary: Now that we've removed BoundTs, we can do a lot less substitution, because we no longer have to substitute BoundTs with GenericTs. Instead, we can just compute using the GenericTs directly. We still need to do substitution in some cases, like PolyT ~> PolyT flows and generic instantiation, but we can entirely remove `check_with_generics` and the associated substitution functions. This has a minor effect on error positioning in some cases.

Reviewed By: panagosg7

Differential Revision: D34018357

fbshipit-source-id: 43b670c2ce0fb879df50e777e4f17c7fe616ebc3