[new-env] remove LocEnvEntryNotFound exception
Summary:
`LocEnvEntryNotFound` used to be raised by `find_var` and then caught in various
functions in new_env.ml. Instead of using exceptions for this logic this diff makes
`find_var` return a `result` type instead, and then handles these values accordingly
in the various callers. This makes it easier to detect cases where the absence of an
entry in the `env_values` map is actually an exception that needs to be addressed.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D34529825

fbshipit-source-id: f30515b28997e565e3704d4a19cb3e3f1da3215c