restrict types that can be normalized in sketchy null checks
Summary:
This new approach to sketchy nulls adds a significant perf regression, which seems to be due to the cost of resolving types post-inference. However, for sketchy nulls, we really only need to know type information shallowly--types like e.g. a class instance or a function can never be involved in sketchy nulls. This diff reduces the perf regression significantly (though not fully) by having the normalization skip a set of such types.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D34513661

fbshipit-source-id: e1b67470602009a4c1952a9b66f52bdbad212be4