Sketchy null checks can be post-inference
Summary:
In preparation for a fix to support the new environment, this diff moves the placement of sketchy null checks to be a post-inference analysis rather than a check during inference. Sketchy null checking shouldn't affect inferred types at all, so this should be a safe change. There are a number of permuted errors, however, where the first nullish element of a union is blamed rather than the second. This is, if anything, and improvement, however.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D34457815

fbshipit-source-id: dce6c59327271e06a53b06f79947f67c5dae148e