Implement rollback function for entities
Summary:
Consider the case where we update entities during transaction A, but we cancel
that transaction without committing. The committed data will stay the same, but
the entity versions and latest data are unchanged.

If we restart another transaction B, once B commits, the writes from A will
become committed as well.

Ideally, we would maintain an invariant that the set of entities written by B is
a superset of those written by A. That is, any incomplete writes from A will
be completed by B.

Mostly, this invariant holds in Flow today, with one exception: priority
rechecks. If we see an unchecked dependency while handling a IDE request, we
will cancel an ongoing recheck (A), schedule a new recheck (B) containing only
the files needed to handle the IDE request, then re-schedule A after B commits.

In this case, committing B will publish the potentially incomplete writes from
A. Instead, we must roll back the entities from A first.

This diff implements a per-entity rollback function. Similar to what we do with
oldify/revive, we will track changed entities and roll them back individually
when a transaction is rolled back.

Reviewed By: mroch

Differential Revision: D33959975

fbshipit-source-id: d244a17fddf87f2bb7546099bcbef2a695afd1ae