Use entities for parsing heap
Summary:
Before this change, rechecks used oldify/revive to update the parsing heap
transactionally. This diff uses entities instead. See D33746701 (https://github.com/facebook/flow/commit/df65dc250180eada9490534078f2d1fa8674d4d1) for background on entities.

Using entities, we no longer need to oldify files or revive them. We no longer
need to keep track of currently oldified files to dispatch reads.

The FileHeap stores pointers to file entities instead of pointers to file data
directly. This indirection means that readers can load the committed data while
mutators can load the latest data.

The first time we see a file, we will create a new entity and store that in the
heap. During a recheck, we might update a file which already has an entity in
the heap, in which case we "advance" that entity to the new version.

Deleted files are removed from the hash table when the transaction commits.

Reviewed By: mroch

Differential Revision: D33725932

fbshipit-source-id: 02a106f77b5cec8d1e44ff4670ad5b6393aa1059