Remove sketchy write_opt function
Summary:
Optional values in shared memory are either 0 (null) or just the original
pointer. There's no tagging and it's not possible to represent `Some None`,
because it would just be 0/null/none.

The `write_opt` API is sketchy because you can write code and expect to get
`Some None`. In general, we just shouldn't provide a `'a addr -> 'a addr opt`
interface of any kind.

I introduced this function because I needed to pass an optional haste module
name when writing files. I have updated these APIs to take a OCaml option
instead.

Reviewed By: mroch

Differential Revision: D33960119

fbshipit-source-id: 40247cdaa3cfeae551b0c002fdddd1305b60a603