Store file kind and name in shared memory
Summary:
Currently we always have a file key, so we didn't need to store this in shared
memory. In some coming changes, this will change so we need to be able to get
this information.

Specifically, we will need to do this when finding a module's provider.
Currently we store a `Modulename.t` -> `File_key.t` mapping (`NameHeap`), but
soon we will store a `file addr` instead. This gives us a way to construct a
`File_key.t` given a `file addr`.

Reviewed By: mroch

Differential Revision: D34309334

fbshipit-source-id: 4314d610ee213c881f518ad1cc2c8036c30c6664