do not unify already bound declare functions
Summary:
This is change in the old environment to match with what the new-env does. In
```
function h(): string { return "" }
declare function h(): bool;
```
flow would raise a [name-already-bound] error, but would also attempt to unify
the types of the two functions, which is unnecessary.

Instead, we keep the [name-already-bound] error, but skip unifying in `unify_declared_fun_type`
when the entry is already a `FunctionBinding`.

Changelog: [fix] do not attempt to unify the type of a "declare" function that comes after a function declaration of the same name ([try-Flow](https://flow.org/try/#0GYVwdgxgLglg9mABACwBQEoBciDOUBOMYA5ogN6L4CmUI+SARA4gL4BQAJlRADYCG1RKEiwEKDNgBGcODwDcQA))

Reviewed By: jbrown215

Differential Revision: D34600522

fbshipit-source-id: 56096d35f2df4581a353cb8f3d9c3ce7a7c23d32