Use lookup loc in refinement lookup for method calls
Summary:
This call to `Refinement.get` incorrectly supplies the location of the entire call instead of the location of the member expression, which makes us miss the refinement on method calls. This diff fixes that, so now heap refinements that are immediately called are read properly.

`Passed: 408, Failed: 60, Skipped: 287, Errored: 0`

Changelog: [Internal]

Reviewed By: gkz

Differential Revision: D34430960

fbshipit-source-id: d1595158cb12e5c884e7bbdfffb54a3f22f2f437