Match new statement.ml switch behavior
Summary:
This diff matches the behavior of the name_resolver to statement.ml as close as possible as implemented in D34006224.

We keep around a copy of the environment after reading the discriminant as the starting case environment. When we finish a case, we add the negation of that case's refinements to the starting case environment and then proceed to the next case until we finish all the cases.

The post-env is computed by taking every env at a switch exit, which we define as each break and the final fallthrough case, and merging them together. Concretely, we merge in all the break envs with the call to `commit_abrupt_completion_matching`, which will merge in all of the `Break None` exits with the current environment. We determine that base environment with the following rules:
1. If there is a default, then we know the switch is exhaustive.
a. If there was a fallthrough_env, then that env + all the break exits covers every env that reaches the post-switch environment. Use the fallthrough_env as a base.
b. If there was no fallthrough_env, then we can use `empty` as a base, since the rest of the envs make up the entire environment. Since there was a default and no fallthrough we know there must at least be one break, so the resulting env will not be empty.
2. If there was no default, use the final case_starting_env, which will have the negation of all refinements, as the starting env. This represents the case in which we never entered the switch statement cases because no case matched. Thus the environments that reach the end in this case are the last case starting env (which less precisely could have just been the pre-switch env, but the negation of refinements is important), all the breaks, and the last fallthrough env if one exists.

We don't need to keep track of total/partial writes like we do in D34006224 because the Val.ts in the env_state are only changed when the vals are different! This is a nice simplification over the previous implementation, in my opinion.

`Passed: 406, Failed: 62, Skipped: 287, Errored: 0`

Changelog: [Internal]

Reviewed By: panagosg7

Differential Revision: D34430764

fbshipit-source-id: 98bb058156e068e49672a3590d09989225d9781c