[new-env] Fix type checking of non-toplevel recursive functions in new_env
Summary:
Changelog: [internal]

## Background

The old and the new environment handles types that should be equal differently. The new env is more systematic: relying on `name_resolver` to compute syntactically an SSA mapping between reads and writes, and can thus can decide type equality in a central place. The old environment, which is designed for non-local type inference, can be more flexible. With old environment can easily setup some open type variables and call `Flow.unify` in whatever place we see fit.

The distinction is the heart of the problem we see in the failing test example (`function/issue-7529.js`). The old environment setups a `knot` as the tvar for the function type of `foo` when visiting its signature and then add `foo: knot` to the scope, which will later be picked by a read of `foo` in the old environment when visiting `foo` in `foo(123)`. The connection between these types are established in the following way:

```
((string) => void) === knot // by visiting function signature
knot === type of `foo` in `foo(123)` // by adding `foo: knot` in the scope
```

The new environment never performs the second step. Instead, it uses the SSA environment to establish that `foo` in `function foo(...) {}` and `foo(123)` should have the same type. However, `statement.ml` never performs a write of `foo` at `function foo(...) {}`, so the type of `foo` is always unconstrained, which causes us to miss the type error.

## Fix

The fix is surprising simple. We create a scope for function name and first `bind_fun` and then `init_fun` in the scope. This is a parallel of how we do it for function declaration, which we `bind_fun` in the initial pass to collect signature and `init_fun` in the second pass to generate constraint. Since the function is not in toplevel position, we can't perform two passes so we do it in one place.

The setup also allows us to completely get rid of the knot approach first introduced in D14835192 (https://github.com/facebook/flow/commit/ace8eaca79170b7be20e3949c4e9525cbb38a15c), so I removed the now useless knot field in function signature.

Reviewed By: jbrown215

Differential Revision: D34627956

fbshipit-source-id: 03f8cfb6a598c8f4f2d0a49cc969232c08797287