catch exceptions in dfind init
Summary:
if the dfind server fails to start, we currently throw an uncaught "End_of_file" exception and exit with an "unknown error" code.

this diff catches those errors and exits with a better error. this is now the same as what happens when watchman fails to init.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D34686398

fbshipit-source-id: 182e60c308a20b5dd6d8d9fcaacf6c78e7ef9f2c