Remove Expensive.t wrapper from get_provider
Summary:
We added this stuff a while back to ensure that expensive shared heap operations
were happening in workers instead of the server process. This was maybe useful
then, but is less relevant these days.

For one, we hit shared memory from the server process in a few places by design:
notably when we handle IDE requests. These calls all have Expensive.warn, which
is true to the spirit of this stuff, but not very useful since there's nothing
we can change.

Secondly, this is not really an expensive operation. We *should* be able to do
this from anywhere. Changes later in this stack will make it less expensive by
moving this data to the new heap.

Reviewed By: mroch

Differential Revision: D34690059

fbshipit-source-id: 213280d5381194e3bc879156387e468748c65ff3