Exact objects have non-strict access with `void` default
Summary:
With this change, if you attempt to access a non-existent property from a union of exact objects, and one or more of those exact objects have that property, but not all do, instead of producing an error we allow the access and infer `void` for the access from the object(s) which do not have the property.

This change is safe as it reflects the realities at runtime - `undefined` is produced when accessing non-existent properties (while retaining the Flow error for accessing a property that does not exist in any case, as this helps with things like refactoring).

E.g.
```
declare var o: {|a: number|} | {|a: number, b: string|};
(o.a: number); // As before
(o.b: string | void); // Now allowed
(o.b: string); // An error
o.XXX; // Still an error
```
For a better idea see the tests in `flow/tests/exact/nonstrict-access.js`, `flow/tests/exact_empty_objects/test.js`, and `flow/tests/exact_empty_objects/defaulting.js`

A big motivation of this change is that when we make the empty object literal an exact empty object (instead of an unsealed object), common patterns like `const o = foo ?? {}; o.prop` will be supported, without special casing the empty object in any way.

This change is only for exact objects. Since accessing a missing property from an inexact object can at best produce `mixed`, I do not believe it is useful to support (this also keeps the error diff down).

I have also kept strict access for a variety of special cased behaviour for which I do not think it would be useful, e.g. for method access (you can't call `undefined`) to keep the error diff down. Some other cases are not implemented yet, e.g. indexed access types - this can be done in the future if desired.

For the above two reasons, while I originally examined re-using `TestPropT`, I decided to add an id to `GetPropT` instead. At least in some of my previous diffs I made the behaviour of `GetPropT` and `TestPropT` more consistent (by adding use ops to TestPropT, and adding support for typo suggestions).

Changelog: [feature] Access from a union of exact objects, on a property that is on one but not all of the objects, now results in the type of that property or `void`, instead of an error

Reviewed By: samwgoldman

Differential Revision: D34408092

fbshipit-source-id: 0319c0921aeae8896b971efd8e5a01e2ffb3b195