fix dfind crash on Windows
Summary:
if a watch root (e.g. an include) doesn't exist, Fsnotify (used by dfind) on Windows would try to raise an ocaml exception from inside a C thread, leading to a crash.

instead, check whether the directory exists outside the thread so we can raise an exception. also, this error is not fatal so we just log and ignore it as we do on darwin in a very similar way (same error message).

Changelog: [fix] Fix a crash when using the default file watcher (dfind) on Windows if an [include] directory doesn't exist.

Reviewed By: samwgoldman

Differential Revision: D34728376

fbshipit-source-id: 1fa4673c751a5843cf7b2dc7af23ed3593a7b89a