[new-env] propagate "predicate" field through various provider analysis structures
Summary:
This change is in service of the next diff. What we'd like to have is a way to determine if
a specific overload is a predicate function (has a `%checks`). AFAIK this information is
not available in the provider information so far. The check that I'd like port is currently
happening in env.ml and uses the old environment scopes.

This diff just propagates the `predicate` field through the provider analysis.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D34658791

fbshipit-source-id: f1d557bb64900a1ff4509c07208d5b0f82a173b6