[new-env] raise EInvalidDeclaration errors
Summary:
For code of the form
```
var z;
z;
```
we raise an [invalid-declaration] error.

This diff ports this to new-env by reusing logic of `promote_non_const` that relies on
provider analysis, and so can be reused for the new-env. (I am assuming this is the
case since `declaration_validity` is under the `env_builder` directory.)

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D34663302

fbshipit-source-id: 4dd4063498bb59c040768f725befd5344b087048