fix uncaught dfind exception
Summary:
when the dfind process dies unexpectedly, two different things notice: 1) the pipe that the monitor reads events from closes (EPIPE), and 2) waitpid on the dfind process unblocks because it exited.

the latter is the better way to handle it because it does a graceful shutdown, instead of `exit` in the middle of nowhere.

Reviewed By: samwgoldman

Differential Revision: D34767804

fbshipit-source-id: 9827e55d310ffd3c0c317c5300a4d5e9bc26f592