Use entities for name heap
Summary:
The name heap stores the provider file for each module. When files change, we
re-pick providers for each dirty module.

Instead of oldify/revive, this diff uses entities. Furthermore, instead of
storing the file key of the provider file, we store a pointer to the file entry
directly.

If a transaction is rolled back, the entities for each dirty module are also
rolled back.

We now create name heap entries during parse. When creating a file entry, we
check if a name heap entry exists. If one does not, we create it and add to the
shared hash table. Adding the name heap entries can race, but the race is
benign. Only one thread will successfully register a name heap entry with the
hash table; data allocated by other threads will be freed by GC.

Querying a module's provider now returns a file address instead of a file key.
This diff adds the necessary APIs to get the file key from a file address if
needed.

Reviewed By: mroch

Differential Revision: D34701147

fbshipit-source-id: 97208492e900354f7addfa96f222546873aa5f26