Use heap to detect changed modules due to changed files
Summary:
When we commit modules, we start with a candidate set of dirty modules. A dirty
module is changed if:
1. It is provided by a different file
2. It is provided by the same file, if that file has changed

For case (2), we used the new_or_changed set. I want to move the commit modules
logic onto workers, and it would be wasteful to send the new_or_changed set to
workers, so this diff implements a different approach.

We can detect that a file is changed by looking at the heap instead. If the
file's parse entity is changed, then the file is changed.

Reviewed By: mroch

Differential Revision: D34691204

fbshipit-source-id: c3aea3a5cbdd78a06915ab39dd40029e5f8cc030