Move NameHeap to Parsing_heaps module
Summary:
NameHeap stores a mapping from "module name" to a provider file. Later in this
stack, I am going to move this logic into a new format. In the new format, we
will create module entries when parsing a file, so it's useful to have the hash
table in the same module where we create file records.

Eventually, I expect the separate "parsing heap" and "module heap" and "context
heap" modules that we have now to kind of boil away, so this is kind of a
temporary shuffling.

Reviewed By: mroch

Differential Revision: D34690404

fbshipit-source-id: aab0b7088db29e4e3f9388bf2ebe65087418e168