[new-env] Correctly report all type binding errors in new env
Summary:
Changelog: [internal]

Right now we are not checking for duplicate bindings for types when we encounter a type identifier in name_resolver. This diff checks for them in `name_resolver#binding_type_identifier`.

The logic is similar to that of `pattern_identifier` and `error_for_assignment_kind`, but the behavior of type binding is significantly different from value binding (no need to havoc, cannot rebind, types are already bind in hoister), so I put it in its own function.

Reviewed By: jbrown215

Differential Revision: D34704006

fbshipit-source-id: 6b5c8cf408b8c154950ac8b2533da3338c95830d