fix LSP restart after multiple config changes
Summary:
Changelog: [fix] Fix a bug where changing a libdef, package.json or .flowconfig multiple times will cause the server to stop and not be restarted.

# details

when the flowconfig changes, the server exits with `Flowconfig_changed` exit code and gets restarted by `flow lsp`. before restarting the server, `flow lsp` reads the flowconfig itself and exits if the version changed; if so, vscode restarts the new `flow lsp` binary.

the same process happens for `Server_out_of_date`, which is used for lib and package.json changes.

for some reason, we only restart the server the first time either of these things happen. the comments say it's to prevent "version battles." but this shouldn't happen because `flow lsp` checks the version before restarting. it's also possible (likely!) that a lib, package.json or flowconfig will change again sometime down the road in the restarted server. we currently don't restart the server in this case and get stuck with "Flow: server is stopped" in vscode.

instead, if the server successfully starts then we should reset the flag to allow the new server to be auto-restarted next time. I suspect we could kill this flag but want to make a less risky fix.

Reviewed By: samwgoldman

Differential Revision: D34642580

fbshipit-source-id: 8bc682729f1baa16702fbbfdfbaac36360e39268