[new-env] Report rebinding errors on existing let declaration
Summary:
Changelog: [internal]

We are missing the case for let binding.

In addition, we need to avoid emitting `ENameAlreadyBound` when we visit the same binding twice. Normally, this is not possible, but it happens in loops when we first visit the LHS in `scout` and then in `visit_guard_and_body`. The environment is indeed correctly reset after `scout` to make the binding being undeclared again, but it's then immediately reset in the following [`havoc_changed_vars`](https://github.com/facebook/flow/blob/main/src/analysis/env_builder/name_resolver.ml#L1880) that changes it to `havoc`, which is not undeclared. This diff changes the reset to happen only if the val is not undeclared, which seems to be in line of [other havoc code](
https://github.com/facebook/flow/blob/29e8eb2e873f5eddea3136b657737393f110db8b/src/analysis/env_builder/name_resolver.ml#L1005-L1015) does.

Reviewed By: jbrown215

Differential Revision: D34729203

fbshipit-source-id: cd6bf4017dbe6cf97f710183e7a5864c797266d0