[lsp] Attach deprecated tag of completion items to supported language clients
Summary:
Changelog: [feature] Show deprecated autocompletion items with strikethrough in IDEs.

This diff starts the support for surfacing information on deprecated symbols in IDEs. This diff completes the easy part: correctly mark autocompletion item as deprecated if they are deprecated. This is relatively easy because we already show documentation along with the completion item name and type, so we can just grab the deprecated info from the parsed jsdoc at the same time.

Reviewed By: gkz

Differential Revision: D34913192

fbshipit-source-id: ca56431cb1044f80401a87e48e7b578c0f84f898