[autocomplete] When autocompleting object keys, change what keys we consider already used and to be excluded
Summary:
Currently when giving autocomplete suggestions for object keys based on upper bounds, we attempt to exclude keys based off of the inferred type of the object literal.

However, this is not ideal as if the upper bound is an object with optional properties, and we supply all required properties, our inferred type will include those optional properties, so we will not get suggestions for them (reported issue).

Additionally, it is a common pattern to spread an object with default values, and then override some of them.

So I have come up with a new approach: we exclude all written down keys, and all keys from spreads which occur after the insertion point (as these would override our insertion). This solves both above problems.

Changelog: [ide] When autocompleting object keys, change what keys we consider already used and to be excluded

Reviewed By: SamChou19815

Differential Revision: D34885413

fbshipit-source-id: 3a4e50debd49e019e7865a94caa60877c968b60d