[parser] Be more tolerant when parsing properties when prepending an object key to an object literal
Summary:
Sometimes when a user is editing code, they want to add a new object property at the beginning or middle of an object literal.
Currently we are not very tolerant when parsing in these situations. For example:
```
const x = {
  foo // <-- inserted here
  bar: 1,
};
```
the above is parsed as two properties:

1) key: `foo`, value: `null`
2) key/value: `1`, shorthand: true (obviously wrong)

With my change below, we recover more gracefully. We parse the above as:

1) key/value: `foo`, shorthand: true
2) key: `bar`, value: `1`

and give a custom error about expected either a colon (for property with init) or comma (for shorthand property)

Changelog: [parser] Be more tolerant when parsing properties when prepending an object key to an object literal

Reviewed By: SamChou19815

Differential Revision: D34885414

fbshipit-source-id: 539207b3132ce4ced38140ad306e1738c91a14a9