Fix calc_old_modules for files which never existed
Summary:
The deleted files in new_or_changed_or_deleted comes from the parsing phase if a
file can not be found. If a deleted file used to exist, we will have a file
object for it in shared memory, posisbly with old parse data, which we use to
find old modules.

However, if Flow never knew about the file before, then we will have no file
object in shared memory. For example, if a file is added and deleted quickly, we
will not yet have created a record for the file, and it will appear as deleted.

To fix, we handle the case where a deleted file has no record. For
calc_old_modules, we can simply ignore such files, since they can not be
registered as former providers of modules.

(Note: this ignores all push blocking failures!)

Reviewed By: pieterv

Differential Revision: D34971375

fbshipit-source-id: 76c2e4524375bb8593bfe84d553b90e97c0362e5