Simpify sharedmem read_opt API
Summary:
Before this change, `read_opt f x` would apply `f` to `x`, like fusing the
`'a opt addr -> 'a addr option` conversion with `Option.map`. Having used this
API for a while now, the fused API is a bit frustrating to use. I often find
myself passing `Fun.id` as `f` which just makes the code more confusing.

This diff removes the fused `Option.map` call and updates callers to simply pass
the result to `Option.map` if that's what we want.

In fact, the stacked diff goes further and removes `'a opt addr` from the public
API entirely. Intead, we just create the `'a addr option` eagerly, so review
this and the next diff together.

Reviewed By: jbrown215

Differential Revision: D34881307

fbshipit-source-id: 6f963c3c066f0ca999bc91ddb24b8db4f300c10f