Remove `_ opt addr` from sharedmem API
Summary:
In shared memory, there are a few places where we have an optional value. We
store the absence of a value as `0` and the presence of one as just the address,
relying on the fact that `0` is never a valid address to a value (and that we
never have nested optional values).

Having used this API for a while, dealing with `_ opt addr` values is more
trouble than it's worth. Almost always we end up calling read_opt, converting to
an `_ addr option` instead. No consumers are so perf sensitive that avoiding
the allocation of the `Some _` value is material.

With this diff, we completely hide the representation of optional values. The
sharedmem API takes `_ option` values as input and returns `_ option` values as
needed.

Reviewed By: jbrown215

Differential Revision: D34881308

fbshipit-source-id: 0a5948292200a422561dc320eecd4ce6366c0639